+++ command/meta_backend_test.go
@@ -1750,7 +1750,7 @@ func TestMetaBackend_configureWithExtra(t *testing.T) {
 
 	// Check the state
 	s := testDataStateRead(t, filepath.Join(DefaultDataDir, backendLocal.DefaultStateFilename))
-	if s.Backend.Hash != cHash {
+	if s.Backend.Hash != uint64(cHash) {
 		t.Fatal("mismatched state and config backend hashes")
 	}
 
@@ -1766,7 +1766,7 @@ func TestMetaBackend_configureWithExtra(t *testing.T) {
 
 	// Check the state
 	s = testDataStateRead(t, filepath.Join(DefaultDataDir, backendLocal.DefaultStateFilename))
-	if s.Backend.Hash != cHash {
+	if s.Backend.Hash != uint64(cHash) {
 		t.Fatal("mismatched state and config backend hashes")
 	}
 }
